books/RAG 마스터 랭체인으로 완성하는 LLM 서비스/rag-master-practice/pyproject.toml
[project]
name = "rag-master-practice"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.3",
    "langsmith>=0.4.46",
    "notebook>=7.5.0",
    "pandas>=2.3.3",
    "pypdf>=6.3.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
